tests/wpt/web-platform-tests/fetch/nosniff/image.html

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
  // Note: images get always sniffed, nosniff doesn't do anything
  var passes = ["", "?type=", "?type=x", "?type=x/x", "?type=image/gif", "?type=image/png", "?type=image/png;blah"]

  passes.forEach(function(urlpart) {
    async_test(function(t) {
      var img = document.createElement("img")
      img.onerror = t.unreached_func("Unexpected error event")
      img.onload = t.step_func_done(function(){
        assert_equals(img.width, 96)
      })
      img.src = "resources/image.py" + urlpart
      document.body.appendChild(img)
    }, "URL query: " + urlpart)
  })
</script>
